+ code refactor + new method
