+++ src/common/index.ts
@@ -1,0 +1,1 @@
+export { common } from './common';
