--- src/index.d.ts
@@ -1,2 +1,0 @@
-export {getVimeoThumbnail} from "./getVimeo"
-export {getYouTubeThumbnail} from "./getYouTube"
