+++ src/index.ts
@@ -1,3 +1,3 @@
 export { getVimeoThumbnail } from './vimeo/index';
 export { getYouTubeThumbnail } from './youtube/index';
-export { common } from "./common/index";
+export { common } from "./common/index"
